// tina-graphql-gateway/packages/tina-graphql/src/primitives/types.ts
export type TinaFieldInner<WithNamespace extends boolean> =
  | ScalarType<WithNamespace>
  | ObjectType<WithNamespace>
  | ReferenceType<WithNamespace>

export type TinaFieldBase = TinaFieldInner<false>
export type TinaFieldEnriched = TinaFieldInner<true>

interface TinaField {
  name: string
  label: string
  description?: string
  required?: boolean
  list?: boolean
  /**
   * Any items passed to the UI field will be passed to the underlying field.
   * NOTE: only serializable values are supported, so functions like `validate`
   * will be ignored.
   */
  ui?: object
}

type ScalarType<WithNamespace extends boolean> = WithNamespace extends true
  ? ScalarTypeWithNamespace
  : ScalarTypeInner
// ...
export type ReferenceType<WithNamespace extends boolean> =
  WithNamespace extends true ? ReferenceTypeWithNamespace : ReferenceTypeInner
export interface ReferenceTypeInner extends TinaField {
  type: 'reference'
  reverseLookup?: { label: string; name: string }
  collections: string[]
}
export interface ReferenceTypeWithNamespace extends TinaField {
  type: 'reference'
  collections: string[]
  reverseLookup?: { label: string; name: string }
  namespace: string[]
}

export type ObjectType<WithNamespace extends boolean> =
  | ObjectTemplates<WithNamespace>
  | ObjectFields<WithNamespace>
// ...
export type Templateable = {
  namespace: string[]
  fields: TinaFieldEnriched[]
  ui?: object
}